+++ src/main/java/shblock/interactivecorporea/common/requestinghalo/HaloCraftingServerHandler.java
@@ -53,9 +53,5 @@ public class HaloCraftingServerHandler {
       return pos.add(centerSlotPos);
     }
 
-//    public boolean tick() {
-//
-//      progress ++;
-//    }
   }
 }